NR: Lesson_4 Homework correction
