+++ src/main/java/student_natalija_rasponomarjova/lesson_4/level_7/Calculator.java
@@ -24,72 +24,5 @@ class Calculator {
     }
 }
 
-class CalculatorTest {
-    public static void main(String[] args) {
-        CalculatorTest calculatorTest = new CalculatorTest();
-        calculatorTest.sumTest();
-        calculatorTest.subTest();
-        calculatorTest.mulTest();
-        calculatorTest.divTest();
-        calculatorTest.isEvenTest1();
-        calculatorTest.isEvenTest2();
-    }
 
-    private void checkResults(boolean condition, String testName) {
-        if (condition) {
-            System.out.println(testName + " = TEST OK");
-        } else {
-            System.out.println(testName + " = TEST FAILED");
-        }
-    }
-
-    public void sumTest() {
-        int firstNumber = 35;
-        int secondNumber = 40;
-        Calculator calculator = new Calculator();
-        int result = calculator.sum(firstNumber, secondNumber);
-
-        checkResults(result == 75, "Sum");
-    }
-
-    public void subTest() {
-        int firstNumber = 35;
-        int secondNumber = 15;
-        Calculator calculator = new Calculator();
-        int result = calculator.sub(firstNumber, secondNumber);
-
-        checkResults(result == 20, "Subtraction");
-    }
-
-    public void mulTest() {
-        int firstNumber = 5;
-        int secondNumber = 5;
-        Calculator calculator = new Calculator();
-        int result = calculator.mul(firstNumber, secondNumber);
-
-        checkResults(result == 25, "Multiplication");
-    }
-
-    public void divTest() {
-        int firstNumber = 40;
-        int secondNumber = 8;
-        Calculator calculator = new Calculator();
-        int result = calculator.div(firstNumber, secondNumber);
-
-        checkResults(result == 5, "Division");
-    }
-
-    public void isEvenTest1() {
-        Calculator calculator = new Calculator();
-        boolean result = calculator.isEven(40);
-
-        checkResults(result, "isEven");
-    }
-
-    public void isEvenTest2() {
-        Calculator calculator = new Calculator();
-        boolean result = calculator.isEven(41);
-        checkResults(!result, "isOdd");
-    }
-}
